split the main prompt file
