+++ .env.template
@@ -3,7 +3,6 @@ GEMINI_API_KEY=
 GPT_MODEL=gpt-5
 FT_MODEL=
 TAVILY_API_KEY=
-NOMINATIM_USER_AGENT=
 CRON_SECRET=
 
 # Neon
