+++ app/api/parrot-chat/route.ts
@@ -7,7 +7,7 @@ import { cookies } from 'next/headers';
 import type { Prisma } from '@prisma/client';
 import { sendError, sendProgress } from '@/lib/progressUtils';
 import prisma from '@/lib/prisma';
-import * as prompts from '@/lib/prompts';
+import * as prompts from '@/lib/prompts/core';
 import { SystemMessage, HumanMessage, AIMessage } from '@langchain/core/messages';
 import { parrotWorkflow } from '@/utils/langChainAgents/mainAgent';
 import { generateConversationName } from '@/utils/generateConversationName';
